backend/requirements.txt
uvicorn
streamlit
requests
loguru
fastapi
uvicorn
loguru
pydantic
tensorflow
pillow
numpy
aiofiles